Move CameraUtils to androidshared
